Minor attempt at making the golang examples a bit better
The instructions in the README assumes the repo will be publicly
accessible on the URL: https://github.com/UKCloud/esc-s3-examples

If that's not the case, we'll have to update the README accordingly.

The only thing I'm not 100% sure of is if the the minio dependecy will
be properly pulled in via `go get`, we are pointing directly at a
directory with Go code in it, but so I believe it will work, but I
have no live repo to test against that does not have any Go code in
the root of the project too.